backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
authlib==1.6.4
httpx==0.28.1
email-validator==2.3.0
python-multipart==0.0.20
itsdangerous==2.2.0
openpyxl==3.1.5
crewai[tools]>=1.13.0
